WIP: query processor refactoring
