+++ libakumuli/queryprocessor.cpp
@@ -45,514 +45,7 @@ Node::NodeType NodeException::get_type() const {
     return type_;
 }
 
-struct RandomSamplingNode : std::enable_shared_from_this<RandomSamplingNode>, Node {
-    const uint32_t                      buffer_size_;
-    std::vector<aku_Sample>             samples_;
-    Rand                                random_;
-    std::shared_ptr<Node>               next_;
 
-    RandomSamplingNode(uint32_t buffer_size, std::shared_ptr<Node> next)
-        : buffer_size_(buffer_size)
-        , next_(next)
-    {
-        samples_.reserve(buffer_size);
-    }
-
-    // Bolt interface
-    virtual NodeType get_type() const {
-        return Node::RandomSampler;
-    }
-
-    bool flush() {
-        auto predicate = [](aku_Sample const& lhs, aku_Sample const& rhs) {
-            auto l = std::make_tuple(lhs.timestamp, lhs.paramid);
-            auto r = std::make_tuple(rhs.timestamp, rhs.paramid);
-            return l < r;
-        };
-
-        std::stable_sort(samples_.begin(), samples_.end(), predicate);
-
-        for(auto const& sample: samples_) {
-            if (next_->put(sample) == false) {
-                return false;
-            }
-        }
-        samples_.clear();
-        return true;
-    }
-
-    virtual void complete() {
-        flush();
-        next_->complete();
-    }
-
-    virtual bool put(const aku_Sample& sample) {
-        if (sample.payload.type == aku_PData::EMPTY) {
-            return flush();
-        } else {
-            if (samples_.size() < buffer_size_) {
-                // Just append new values
-                samples_.push_back(sample);
-            } else {
-                // Flip a coin
-                uint32_t ix = random_() % samples_.size();
-                if (ix < buffer_size_) {
-                    samples_.at(ix) = sample;
-                }
-            }
-        }
-        return true;
-    }
-
-    void set_error(aku_Status status) {
-        next_->set_error(status);
-    }
-};
-
-
-/** Filter ids using predicate.
-  * Predicate is an unary functor that accepts parameter of type aku_ParamId - fun(aku_ParamId) -> bool.
-  */
-template<class Predicate>
-struct FilterByIdNode : std::enable_shared_from_this<FilterByIdNode<Predicate>>, Node {
-    //! Id matching predicate
-    Predicate op_;
-    std::shared_ptr<Node> next_;
-
-    FilterByIdNode(Predicate pred, std::shared_ptr<Node> next)
-        : op_(pred)
-        , next_(next)
-    {
-    }
-
-    // Bolt interface
-    virtual void complete() {
-        next_->complete();
-    }
-
-    virtual bool put(const aku_Sample& sample) {
-        if (sample.payload.type == aku_PData::EMPTY) {
-            return next_->put(sample);
-        }
-        return op_(sample.paramid) ? next_->put(sample) : true;
-    }
-
-    void set_error(aku_Status status) {
-        if (!next_) {
-            AKU_PANIC("bad query processor node, next not set");
-        }
-        next_->set_error(status);
-    }
-
-    virtual NodeType get_type() const {
-        return NodeType::FilterById;
-    }
-};
-
-
-//! Generic piecewise aggregate approximation
-template<class State>
-struct PAA : Node {
-    std::shared_ptr<Node> next_;
-    std::unordered_map<aku_ParamId, State> counters_;
-
-    PAA(std::shared_ptr<Node> next)
-        : next_(next)
-    {
-    }
-
-    bool average_samples(aku_Timestamp ts) {
-        for (auto& pair: counters_) {
-            State& state = pair.second;
-            if (state.ready()) {
-                aku_Sample sample;
-                sample.paramid = pair.first;
-                sample.payload.float64 = state.value();
-                sample.payload.type = AKU_PAYLOAD_FLOAT;
-                sample.payload.size = sizeof(aku_Sample);
-                sample.timestamp = ts;
-                state.reset();
-                if (!next_->put(sample)) {
-                    return false;
-                }
-            }
-        }
-        if (!next_->put(EMPTY_SAMPLE)) {
-            return false;
-        }
-        return true;
-    }
-
-    virtual void complete() {
-        next_->complete();
-    }
-
-    virtual bool put(const aku_Sample &sample) {
-        // ignore BLOBs
-        if (sample.payload.type == aku_PData::EMPTY) {
-            if (!average_samples(sample.timestamp)) {
-                return false;
-            }
-        } else {
-            auto& state = counters_[sample.paramid];
-            state.add(sample);
-        }
-        return true;
-    }
-
-    virtual void set_error(aku_Status status) {
-        next_->set_error(status);
-    }
-
-    virtual NodeType get_type() const {
-        return Node::Resampler;
-    }
-};
-
-struct MeanCounter {
-    double acc = 0;
-    size_t num = 0;
-
-    void reset() {
-        acc = 0;
-        num = 0;
-    }
-
-    double value() const {
-        return acc/num;
-    }
-
-    bool ready() const {
-        return num != 0;
-    }
-
-    void add(aku_Sample const& value) {
-        acc += value.payload.float64;
-        num++;
-    }
-};
-
-struct MeanPAA : PAA<MeanCounter> {
-
-    MeanPAA(std::shared_ptr<Node> next)
-        : PAA<MeanCounter>(next)
-    {
-    }
-
-    virtual NodeType get_type() const override {
-        return Node::MovingAverage;
-    }
-};
-
-struct MedianCounter {
-    mutable std::vector<double> acc;
-
-    void reset() {
-        std::vector<double> tmp;
-        std::swap(tmp, acc);
-    }
-
-    double value() const {
-        if (acc.empty()) {
-            AKU_PANIC("`ready` should be called first");
-        }
-        if (acc.size() < 2) {
-            return acc.at(0);
-        }
-        auto middle = acc.begin();
-        std::advance(middle, acc.size() / 2);
-        std::partial_sort(acc.begin(), middle, acc.end());
-        return *middle;
-    }
-
-    bool ready() const {
-        return !acc.empty();
-    }
-
-    void add(aku_Sample const& value) {
-        acc.push_back(value.payload.float64);
-    }
-};
-
-struct MedianPAA : PAA<MedianCounter> {
-
-    MedianPAA(std::shared_ptr<Node> next)
-        : PAA<MedianCounter>(next)
-    {
-    }
-
-    virtual NodeType get_type() const override {
-        return Node::MovingMedian;
-    }
-};
-
-template<bool weighted>
-struct SpaceSaver : Node {
-    std::shared_ptr<Node> next_;
-
-    struct Item {
-        double count;
-        double error;
-    };
-
-    std::unordered_map<aku_ParamId, Item> counters_;
-    //! Capacity
-    double N;
-    const size_t M;
-    const double P;
-
-    /** C-tor.
-      * @param error is a allowed error value between 0 and 1
-      * @param portion is a frequency (or weight) portion that we interested in
-      * Object should report all items wich frequencies is greater then (portion-error)*N
-      * where N is a number of elements (or total weight of all items in a stream).
-      */
-    SpaceSaver(double error, double portion, std::shared_ptr<Node> next)
-        : next_(next)
-        , N(0)
-        , M(ceil(1.0/error))
-        , P(portion)  // between 0 and 1
-    {
-        assert(P >= 0.0);
-        assert(P <= 1.0);
-    }
-
-    bool count() {
-        std::vector<aku_Sample> samples;
-        auto support = N*P;
-        for (auto it: counters_) {
-            auto estimate = it.second.count - it.second.error;
-            if (support < estimate) {
-                aku_Sample s;
-                s.paramid = it.first;
-                s.payload.type = aku_PData::PARAMID_BIT|aku_PData::FLOAT_BIT;
-                s.payload.float64 = it.second.count;
-                s.payload.size = sizeof(aku_Sample);
-                samples.push_back(s);
-            }
-        }
-        std::sort(samples.begin(), samples.end(), [](const aku_Sample& lhs, const aku_Sample& rhs) {
-            return lhs.payload.float64 > rhs.payload.float64;
-        });
-        for (const auto& s: samples) {
-            if (!next_->put(s)) {
-                return false;
-            }
-        }
-        counters_.clear();
-        return true;
-    }
-
-    virtual void complete() {
-        count();
-        next_->complete();
-    }
-
-    virtual bool put(const aku_Sample &sample) {
-        if (sample.payload.type == aku_PData::EMPTY) {
-            return count();
-        }
-        if (weighted) {
-            if ((sample.payload.type&aku_PData::FLOAT_BIT) == 0) {
-                return true;
-            }
-        }
-        auto id = sample.paramid;
-        auto weight = weighted ? sample.payload.float64 : 1.0;
-        auto it = counters_.find(id);
-        if (it == counters_.end()) {
-            // new element
-            double count = weight;
-            double error = 0;
-            if (counters_.size() == M) {
-                // remove element with smallest count
-                size_t min = std::numeric_limits<size_t>::max();
-                auto min_iter = it;
-                for (auto i = counters_.begin(); i != counters_.end(); i++) {
-                    if (i->second.count < min) {
-                        min_iter = i;
-                        min = i->second.count;
-                    }
-                }
-                counters_.erase(min_iter);
-                count += min;
-                error  = min;
-            }
-            counters_[id] = { count, error };
-        } else {
-            // increment
-            it->second.count += weight;
-        }
-        N += weight;
-        return true;
-    }
-
-    virtual void set_error(aku_Status status) {
-        next_->set_error(status);
-    }
-
-    virtual NodeType get_type() const {
-        return Node::SpaceSaver;
-    }
-};
-
-
-struct AnomalyDetector : Node {
-    typedef std::unique_ptr<AnomalyDetectorIface> PDetector;
-
-    enum FcastMethod {
-        SMA,
-        SMA_SKETCH,
-        EWMA,
-        EWMA_SKETCH,
-        DOUBLE_EXP_SMOOTHING,
-        DOUBLE_EXP_SMOOTHING_SKETCH,
-        HOLT_WINTERS,
-        HOLT_WINTERS_SKETCH,
-    };
-
-    std::shared_ptr<Node> next_;
-    PDetector detector_;
-
-    AnomalyDetector(uint32_t nhashes,
-                    uint32_t bits,
-                    double   threshold,
-                    double   alpha,
-                    double   beta,
-                    double   gamma,
-                    int      period,
-                    FcastMethod method,
-                    std::shared_ptr<Node> next)
-        : next_(next)
-    {
-        try {
-            switch(method) {
-            case SMA:
-                detector_ = AnomalyDetectorUtil::create_precise_sma(threshold, period);
-                break;
-            case SMA_SKETCH:
-                detector_ = AnomalyDetectorUtil::create_approx_sma(nhashes, 1 << bits, threshold, period);
-                break;
-            case EWMA:
-                detector_ = AnomalyDetectorUtil::create_precise_ewma(threshold, alpha);
-                break;
-            case EWMA_SKETCH:
-                detector_ = AnomalyDetectorUtil::create_approx_ewma(nhashes, 1 << bits, threshold, alpha);
-                break;
-            case DOUBLE_EXP_SMOOTHING:
-                detector_ = AnomalyDetectorUtil::create_precise_double_exp_smoothing(threshold, alpha, gamma);
-                break;
-            case DOUBLE_EXP_SMOOTHING_SKETCH:
-                detector_ = AnomalyDetectorUtil::create_approx_double_exp_smoothing(nhashes, 1 << bits, threshold, alpha, gamma);
-                break;
-            case HOLT_WINTERS:
-                detector_ = AnomalyDetectorUtil::create_precise_holt_winters(threshold, alpha, beta, gamma, period);
-                break;
-            case HOLT_WINTERS_SKETCH:
-                detector_ = AnomalyDetectorUtil::create_approx_holt_winters(nhashes, 1 << bits, threshold, alpha, beta, gamma, period);
-                break;
-            default:
-                std::logic_error err("AnomalyDetector building error");  // invalid use of the constructor
-                BOOST_THROW_EXCEPTION(err);
-            }
-        } catch (...) {
-            // std::cout << boost::current_exception_diagnostic_information() << std::endl;
-            throw;
-        }
-    }
-
-    virtual void complete() {
-        next_->complete();
-    }
-
-    virtual bool put(const aku_Sample &sample) {
-        if (sample.payload.type == aku_PData::EMPTY) {
-            detector_->move_sliding_window();
-            return next_->put(sample);
-        } else if (sample.payload.type & aku_PData::FLOAT_BIT) {
-            /*
-            if (sample.payload.float64 < 0.0) {
-                set_error(AKU_EANOMALY_NEG_VAL);
-                return false;
-            }
-            */
-            detector_->add(sample.paramid, sample.payload.float64);
-            if (detector_->is_anomaly_candidate(sample.paramid)) {
-                aku_Sample anomaly = sample;
-                anomaly.payload.type |= aku_PData::URGENT;
-                return next_->put(anomaly);
-            }
-        }
-        // Ignore BLOBs
-        return true;
-    }
-
-    virtual void set_error(aku_Status status) {
-        next_->set_error(status);
-    }
-
-    virtual NodeType get_type() const {
-        return Node::AnomalyDetector;
-    }
-};
-
-
-//                      //
-//      SAX Encoder     //
-//                      //
-
-struct SAXNode : Node {
-
-    std::shared_ptr<Node> next_;
-    std::unordered_map<aku_ParamId, SAX::SAXEncoder> encoders_;
-    int window_width_;
-    int alphabet_size_;
-    bool disable_value_;
-
-    SAXNode(int alphabet_size, int window_width, bool disable_original_value, std::shared_ptr<Node> next)
-        : next_(next)
-        , window_width_(window_width)
-        , alphabet_size_(alphabet_size)
-        , disable_value_(disable_original_value)
-    {
-    }
-
-    void complete() {
-        next_->complete();
-    }
-
-    bool put(const aku_Sample &sample) {
-        if (sample.payload.type != aku_PData::EMPTY) {
-            SAX::SAXWord word;
-            auto it = encoders_.find(sample.paramid);
-            if (it == encoders_.end()) {
-                encoders_[sample.paramid] = SAX::SAXEncoder(alphabet_size_, window_width_);
-                it = encoders_.find(sample.paramid);
-            }
-            size_t ssize = sizeof(aku_Sample) + window_width_;
-            void* ptr = alloca(ssize);
-            aku_Sample* psample = new (ptr) aku_Sample();
-            *psample = sample;
-            psample->payload.size = ssize;
-            psample->payload.type |= aku_PData::SAX_WORD;
-            if (disable_value_) {
-                psample->payload.type &= ~aku_PData::FLOAT_BIT;
-            }
-            if (it->second.encode(sample.payload.float64, psample->payload.data, window_width_)) {
-                return next_->put(*psample);
-            }
-        }
-        return true;
-    }
-
-    void set_error(aku_Status status) {
-        next_->set_error(status);
-    }
-
-    NodeType get_type() const {
-        return Node::SAX;
-    }
-};
 
 
 //                                   //
